% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedshiftSQL.r
\name{RedshiftSQL}
\alias{RedshiftSQL}
\title{Instantiate a Redshift client}
\usage{
RedshiftSQL()
}
\description{
This function creates and initializes a PostgreSQL client with class
RedshiftSQLDriver which is simply a superclass of PostgreSQLDriver
}
\examples{
\dontrun{
con <- dbConnect(RedshiftSQL(), user="u", password = "p", host="h", dbname="n", port = "5439")
query <- dbSendQuery(con, "SELECT * FROM table")
}
}

